A lot of commits for new generation
Setting max hidden apps to 55

Now that we know this is not the culprit of the lag lets use a value that is in a *sweet spot* between 40 and 60 leaning towards 60. (as steve K said all of this shit is arbitrary but In my testing 60 runs great)

Revert "Handle nested folder case"

This reverts commit 2e3e0c3a493aec4bea60d4195a4290e1bd08001b.

Revert "Decrease lag generated by autorotation on"

This reverts commit 47f58dcc33f98dc897ccd6229d20fd82ddf94c70.

gps: fixed logic for finding the best package

Fixed the logic for finding the best package,
as it was always picking the last package in
the list as the best (regardless of which package
had the highest version) if there were multiple
packages with higher versions that the current.

Change-Id: Ibd61f0a050d65feb8bb1584f59b0a11974333af5

Clear Recents: Free fall of layout

Signed-off-by: Daniel <daxx4182@gmail.com>

It helps to catch the right exception.

Bug: 7238149
Change-Id: I7b60ec44d1d1d5ece8a871cb089f811b611f4a4a

Allow theming of lockscreen

Change-Id: I10799482e0b393746b4c9818f94862109fac6006

Remove the fixed-length buffer in DdmHandleNativeHeap.getLeakInfo.

It was possible to overrun this buffer, and even if you did fill
the buffer in a single read and exit the loop, the "string" still
wouldn't be terminated.

Bug: 8468088
Change-Id: Ia19f4d26dfd79e7b63807a7ec9540b932163d246

Two REMOVE_PROVIDER messages caused by race condition.

Fix a bug in unstable ContentProvider.
IllegalStateException: ref counts can't go to zero here: stable=0 unstable=0
IllegalStateException: unstable count < 0: -1

There is a race between main thread and background database thread. Main thread
is responsible for handling the REMOVE_PROVIDER message. Database thread starts
insert or query request again and again. acquireProvider in db thread will often
snatch provider from the jaws of death, sometime it fails to remove REMOVE_PROVIDER
which is already fired out from MessageQueue. But completeRemoveProvider in main
thread gets suspended when trying to execute the critical section. If db thread
released the provider before main thread resumes the execution, then two
REMOVE_PROVIDER messages will be executed.

Change-Id: I8588aa1d1a8bc444dcd2adf6f8bc3f055cebbdc4
Signed-off-by: Guobin Zhang <guobin.zhang@intel.com>

Miragefy

Change-Id: I595419d60d4a3937a4b0a8ed666193031fa7dfd0